refactor: move mq rest exceptions to module
